Move forced activity language mode to shared languages proto
