+++ app/src/main/java/org/oppia/android/app/translation/AppLanguageWatcherMixin.kt
@@ -2,6 +2,7 @@ package org.oppia.android.app.translation
 
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Observer
+import org.oppia.android.app.model.ForcedActivityLanguageMode
 import org.oppia.android.domain.locale.LocaleController
 import org.oppia.android.domain.oppialogger.OppiaLogger
 import org.oppia.android.domain.profile.ProfileManagementController
@@ -10,18 +11,6 @@ import org.oppia.android.util.data.AsyncResult
 import org.oppia.android.util.data.DataProviders.Companion.toLiveData
 import org.oppia.android.util.locale.OppiaLocale
 import javax.inject.Inject
-
-/** Defines how an activity should resolve its locale. */
-enum class ForcedActivityLanguageMode {
-  /** Use the user's selected app language, or the system default if no app language is set. */
-  USE_APP_LANGUAGE,
-
-  /** Use the system's default language, ignoring the user's app language selection. */
-  USE_SYSTEM_LANGUAGE,
-
-  /** Force English locale, regardless of system or app language settings. */
-  USE_ENGLISH
-}
 
 /**
  * Activity mixin for automatically monitoring & recreating the activity whenever the current app
@@ -86,7 +75,9 @@ class AppLanguageWatcherMixin @Inject constructor(
         translationController.getSystemLanguageLocale()
       ForcedActivityLanguageMode.USE_ENGLISH ->
         translationController.getEnglishLocale()
-      ForcedActivityLanguageMode.USE_APP_LANGUAGE -> {
+      ForcedActivityLanguageMode.USE_APP_LANGUAGE,
+      ForcedActivityLanguageMode.FORCED_ACTIVITY_LANGUAGE_MODE_UNSPECIFIED,
+      null -> {
         if (currentUserProfileId == null) {
           translationController.getSystemLanguageLocale()
         } else {
