+++ app/src/test/java/org/oppia/android/app/translation/AppLanguageWatcherMixinTest.kt
@@ -28,6 +28,7 @@ import org.oppia.android.app.application.testing.TestingBuildFlavorModule
 import org.oppia.android.app.devoptions.DeveloperOptionsModule
 import org.oppia.android.app.devoptions.DeveloperOptionsStarterModule
 import org.oppia.android.app.model.AppLanguageSelection
+import org.oppia.android.app.model.ForcedActivityLanguageMode
 import org.oppia.android.app.model.LegacyProfileId
 import org.oppia.android.app.model.OppiaLanguage
 import org.oppia.android.app.model.OppiaLanguage.BRAZILIAN_PORTUGUESE
@@ -263,11 +264,12 @@ class AppLanguageWatcherMixinTest {
 
   @Test
   fun testForcedActivityLanguageMode_values_containsExpectedModes() {
-    assertThat(ForcedActivityLanguageMode.values().asList()).containsExactly(
+    assertThat(ForcedActivityLanguageMode.values().asList()).containsAtLeast(
+      ForcedActivityLanguageMode.FORCED_ACTIVITY_LANGUAGE_MODE_UNSPECIFIED,
       ForcedActivityLanguageMode.USE_APP_LANGUAGE,
       ForcedActivityLanguageMode.USE_SYSTEM_LANGUAGE,
       ForcedActivityLanguageMode.USE_ENGLISH
-    ).inOrder()
+    )
   }
 
   private fun updateAppLanguageTo(language: OppiaLanguage) {
